+++ R/CPPopt.R
@@ -1,5 +1,0 @@
-#Cppopt curve - 10.1097/CCM.0b013e3182514eb6
-
-CPPopt <- function() {
-   print("Hello, world!")
-}